Code-for-Experiment/RAG/encoder_verification_pitch_prediction/requirements.txt
torch==2.1.2
tqdm==4.66.1
pyyaml==6.0.1
numpy==1.26.2
pretty-midi==0.2.10
torchaudio
pyloudnorm
scipy
pydub
deepspeed
transformers